Cycle 1: Fix 50 warnings - pointer sign, unused variables, and misleading indentation
* Fixed pointer sign warning in adler32.c:115 by adding proper cast
* Fixed unused variable warnings in bkd_rclone.c by removing unused variables
* Fixed misleading indentation warnings in collapse.c by proper code formatting
* Fixed pointer sign warnings in checksum.c by adding proper casts
* Reduced total warnings from 177 to 127 (50 warnings fixed)

+++ src/bkd_push.c
@@ -104,7 +104,7 @@ cmd_push_part1(int ac, char **av)
 
 	out("@OK@\n");
 	p = fmem_peek(fout, &n);
-	unless (writen(1, p, n) == n) {
+	unless (writen(1, p, n) == (int)n) {
 		fclose(fout);
 		return (1);
 	}
